use 64 bit offsets
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@65206 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/rc/hdd.c
@@ -420,7 +420,7 @@ rewrite:;
 	}
 
 	fseek(f_kernel, 0, SEEK_END);
-	size_t kernellen = ftello(f_kernel);
+	off_t kernellen = ftello(f_kernel);
 	rewind(f_kernel);
 
 	if (STORE32_LE(trx.flag_version) & TRX_NO_HEADER)
